Added plots, fixed logical structure for which trees to abalyse for which datasets

+++ code/4_DataAnalysis.R
@@ -8,7 +8,7 @@
 ##### Step 1: Open packages #####
 library(ape)
 library(distory)
-#library(ggplot2)
+library(ggplot2)
 #library(adegenet)
 #library(treespace)
 #library(phangorn) # using phangorn to get the KF.dist, RF.dist, wRF.dist, nNodes, and patristic methods for summarising trees as vectors 
@@ -46,7 +46,7 @@ tests_to_run <- list("Vanderpool2020" = c("allTests", "PHI", "maxchi", "geneconv
                      "Strassert2021" = c("PHI", "maxchi"),
                      "1KP" = c("PHI", "maxchi"))
 
-datasets_to_identify_distinct_edges <- c("Pease2016", "Vanderpool2020")
+datasets_to_identify_distinct_edges <- c("Pease2016", "Vanderpool2020", "Strassert2021", "1KP")
 
 
 ##### Step 3: Source function files and prepare variables for analysis #####
@@ -60,7 +60,7 @@ names(alignment_dir) <- input_names
 
 
 
-##### Step 4: Compare the posterior probabilities/ bootstraps of the trees #####
+##### Step 4: Extract the posterior probabilities/ bootstraps of the trees #####
 node_output_dir <- paste0(output_dir, "node_comparisons/")
 if (dir.exists(node_output_dir) == FALSE){
   dir.create(node_output_dir)
@@ -75,38 +75,58 @@ for (dataset in datasets_to_identify_distinct_edges){
   # Iterate through each test and identify that information
   for (test in dataset_tests){
     print(paste0("Processing ", dataset, ": ", test))
-    ## IQ-Tree trees: Create dataframe detailing differences in posterior probabilities between the two trees
-    # Get the list of trees estimated in IQ-Tree for this dataset
-    test_trees <- grep(test, all_files, value = TRUE)
-    iq_trees <- grep(".contree", test_trees, value = TRUE)
-    # Make the full filepaths for each of the three trees (test pass, test fail, and no test)
-    none_tree_file <- paste0(dataset_tree_dir, grep(".contree", grep("NoTest", all_files, value = TRUE), value = TRUE))
-    pass_tree_file <- paste0(dataset_tree_dir, grep("pass", iq_trees, value = TRUE))
-    fail_tree_file <- paste0(dataset_tree_dir, grep("fail", iq_trees, value = TRUE))
-    # Create dataframes
-    # Want to collect the information about splits present in one tree but not the other (i.e. in T_all,pass vs T_None)
-    test_df_pass_iq <- compare.distinct.edges.of.two.trees(tree_file_1 = pass_tree_file, tree_file_2 = none_tree_file, tree1_name = "Pass", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "BS")
-    test_df_fail_iq <- compare.distinct.edges.of.two.trees(tree_file_1 = fail_tree_file, tree_file_2 = none_tree_file, tree1_name = "Fail", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "BS")
     
-    ## ASTRAL trees: Create dataframe detailing differences in posterior probabilities between the two trees
-    # Get the list of trees estimated in ASTRAL for this dataset
-    test_trees <- grep(test, all_files, value = TRUE)
-    astral_trees <- grep(".tre", grep(".ASTRAL", test_trees, value = TRUE), value = TRUE)
-    pass_tree_file <- paste0(dataset_tree_dir, grep("pass", astral_trees, value = TRUE))
-    fail_tree_file <- paste0(dataset_tree_dir, grep("fail", astral_trees, value = TRUE))
-    none_tree_file <- paste0(dataset_tree_dir, grep("NoTest", grep(".tre", grep(".ASTRAL", all_files, value = TRUE), value = TRUE), value = TRUE))
-    # Create dataframes
-    # Want to collect the information about splits present in one tree but not the other (i.e. in T_all,pass vs T_None)
-    test_df_pass_astral <- compare.distinct.edges.of.two.trees(tree_file_1 = pass_tree_file, tree_file_2 = none_tree_file, tree1_name = "Pass", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "PP")
-    test_df_fail_astral <- compare.distinct.edges.of.two.trees(tree_file_1 = fail_tree_file, tree_file_2 = none_tree_file, tree1_name = "Fail", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "PP")
-    
-    # Combine all four dataframes into one
-    test_df <- rbind(test_df_pass_iq, test_df_fail_iq, test_df_pass_astral, test_df_fail_astral)
-    # Save dataset
     test_df_filename <- paste0(node_output_dir, dataset, "_", test, "_ExtractDistinctEdges.csv")
-    write.csv(test_df, file = test_df_filename)
+    
+    # If results csv does not exist, calculate results
+    if (file.exists(test_df_filename) == FALSE){
+      
+      # Run IQ-Tree for shallow datasets with completed IQ-Tree species trees
+      if (dataset == "Vanderpool2020" | dataset == "Pease2016"){
+        print("Compare IQ-Tree trees")
+        ## IQ-Tree trees: Create dataframe detailing differences in posterior probabilities between the two trees
+        # Get the list of trees estimated in IQ-Tree for this dataset
+        test_trees <- grep(test, all_files, value = TRUE)
+        iq_trees <- grep(".contree", test_trees, value = TRUE)
+        # Make the full filepaths for each of the three trees (test pass, test fail, and no test)
+        none_tree_file <- paste0(dataset_tree_dir, grep(".contree", grep("NoTest", all_files, value = TRUE), value = TRUE))
+        pass_tree_file <- paste0(dataset_tree_dir, grep("pass", iq_trees, value = TRUE))
+        fail_tree_file <- paste0(dataset_tree_dir, grep("fail", iq_trees, value = TRUE))
+        # Create dataframes
+        # Want to collect the information about splits present in one tree but not the other (i.e. in T_all,pass vs T_None)
+        test_df_pass_iq <- compare.distinct.edges.of.two.trees(tree_file_1 = pass_tree_file, tree_file_2 = none_tree_file, tree1_name = "Pass", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "BS")
+        test_df_fail_iq <- compare.distinct.edges.of.two.trees(tree_file_1 = fail_tree_file, tree_file_2 = none_tree_file, tree1_name = "Fail", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "BS")
+      }
+      
+      print("Compare ASTRAL trees")
+      ## ASTRAL trees: Create dataframe detailing differences in posterior probabilities between the two trees
+      # Get the list of trees estimated in ASTRAL for this dataset
+      test_trees <- grep(test, all_files, value = TRUE)
+      astral_trees <- grep(".tre", grep(".ASTRAL", test_trees, value = TRUE), value = TRUE)
+      pass_tree_file <- paste0(dataset_tree_dir, grep("pass", astral_trees, value = TRUE))
+      none_tree_file <- paste0(dataset_tree_dir, grep("NoTest", grep(".tre", grep(".ASTRAL", all_files, value = TRUE), value = TRUE), value = TRUE))
+      # Create dataframes
+      # Want to collect the information about splits present in one tree but not the other (i.e. in T_all,pass vs T_None)
+      test_df_pass_astral <- compare.distinct.edges.of.two.trees(tree_file_1 = pass_tree_file, tree_file_2 = none_tree_file, tree1_name = "Pass", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "PP")
+      # Run fail tree for shallow datasets only
+      if (dataset == "Vanderpool2020" | dataset == "Pease2016"){
+        fail_tree_file <- paste0(dataset_tree_dir, grep("fail", astral_trees, value = TRUE))
+        test_df_fail_astral <- compare.distinct.edges.of.two.trees(tree_file_1 = fail_tree_file, tree_file_2 = none_tree_file, tree1_name = "Fail", tree2_name = "None", test_name = test, dataset_name = dataset, support_value_type_name = "PP")
+      }
+      
+      print("Saving dataframe")
+      # Combine all four dataframes into one
+      if (dataset == "Vanderpool2020" | dataset == "Pease2016"){
+        test_df <- rbind(test_df_pass_iq, test_df_fail_iq, test_df_pass_astral, test_df_fail_astral)
+      } else {
+        test_df <- rbind(test_df_pass_astral)
+      }
+      # Save dataset
+      write.csv(test_df, file = test_df_filename)
+    }
   }
 }
+
 
 # Collate all dataframes
 all_csvs <- list.files(node_output_dir)
@@ -117,5 +137,109 @@ node_df <- as.data.frame(do.call(rbind, all_csv_dfs))
 node_df_filename <- paste0(node_output_dir, "Collated_ExtractDistinctEdges.csv")
 write.csv(node_df, file = node_df_filename)
 
+
+
+##### Step 5: Compare the posterior probabilities/ bootstraps of the trees #####
+#### Create new columns to facilitate plotting ####
+# Create a new factored dataset columns for nice plotting
+node_df$dataset_fac <- factor(node_df$dataset, levels = c("Vanderpool2020", "Pease2016", "Strassert2021", "1KP"), 
+                              labels = c("Vanderpool et al. (2020)", "Pease et al. (2016)", "Strassert et al. (2020)", "1000 Plants"), 
+                              ordered = TRUE)
+node_df$edge_type_fac <- factor(node_df$edge_type, levels = c("Congruent", "Conflicting"), ordered = TRUE)
+node_df$test_fac <- factor(node_df$test, levels = c("PHI", "maxchi", "geneconv", "allTests"), labels = c("PHI", "MaxChi", "GENECONV", "All tests"), ordered = TRUE)
+
+# Create a new comparison tree column 
+pass_inds <- which(node_df$tree1 == "Pass" | node_df$tree2 == "Pass")
+fail_inds <- which(node_df$tree1 == "Fail" | node_df$tree2 == "Fail")
+node_df$comparison_tree <- NA
+node_df$comparison_tree[pass_inds] <- "Pass"
+node_df$comparison_tree[fail_inds] <- "Fail"
+# Create a column combining the comparison tree and the edge type columns
+node_df$boxplot_groups <- paste0(node_df$comparison_tree, ",\n ", node_df$edge_type_fac)
+node_df$boxplot_fac <- factor(node_df$boxplot_groups, levels = c("Pass,\n Congruent", "Pass,\n Conflicting", "Fail,\n Congruent", "Fail,\n Conflicting"), ordered = TRUE)
+
+#### Plot the shallow datasets ####
+# Separate out the two shallow datasets
+shallow_df <- node_df[(node_df$dataset == "Vanderpool2020" | node_df$dataset == "Pease2016"),]
+
+# Separate into ASTRAL and IQ-Tree data frames
+pp_df <- shallow_df[(shallow_df$support_value_type == "PP"),]
+bs_df <- shallow_df[(shallow_df$support_value_type == "BS"),]
+
+# Make a nice plot of the shallow datasets, faceted by test and dataset
+# For ASTRAL trees (with posterior probabilities)
+p = ggplot(data = pp_df, aes(x = boxplot_fac, y = support_value)) + geom_boxplot() +
+  facet_grid(dataset_fac ~ test_fac) + 
+  theme_bw() +
+  labs(title = "Coalescent trees: posterior probability of conflicted and congruent branches") +
+  scale_x_discrete(name = "Tree and edge type") +
+  scale_y_continuous(name = "Posterior Probability", breaks = seq(0,1,0.1),  labels = seq(0,1,0.1), minor_breaks = seq(0,1,0.05), limits = c(0,1)) +
+  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 90, hjust = 1))
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_ASTRAL_posteriorProbability_conflicting_branches.pdf"), plot = p, device = "pdf")
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_ASTRAL_posteriorProbability_conflicting_branches.png"), plot = p, device = "png")
+
+# For ASTRAL trees (with branch lengths)
+p = ggplot(data = pp_df, aes(x = boxplot_fac, y = edge_length)) + geom_boxplot() +
+  facet_grid(dataset_fac ~ test_fac) + 
+  theme_bw() +
+  labs(title = "Coalescent trees: branch length of conflicted and congruent branches") +
+  scale_x_discrete(name = "Tree and edge type") +
+  scale_y_continuous(name = "Branch length (coalescent units)", breaks = seq(0,6,2),  labels = seq(0,6,2), minor_breaks = seq(0,7,0.5), limits = c(0,7)) +
+  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 90, hjust = 1))
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_ASTRAL_branchLength_conflicting_branches.pdf"), plot = p, device = "pdf")
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_ASTRAL_branchLength_conflicting_branches.png"), plot = p, device = "png")
+
+# For IQ-Tree trees (with ultrafast bootstrap suport values)
+p = ggplot(data = bs_df, aes(x = boxplot_fac, y = support_value)) + geom_boxplot() +
+  facet_grid(dataset_fac ~ test_fac) + 
+  theme_bw() +
+  labs(title = "Concatenated trees: UFB value for conflicted and congruent branches") +
+  scale_x_discrete(name = "Tree and edge type") +
+  scale_y_continuous(name = "Ultrafast bootstrap support value", breaks = seq(0,100,20), labels = seq(0,100,20), minor_breaks = seq(0,100,10), limits = c(0,100)) +
+  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 90, hjust = 1))
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_IQTree_UFBootstraps_conflicting_branches.pdf"), plot = p, device = "pdf")
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_IQTree_UFBootstraps_conflicting_branches.png"), plot = p, device = "png")
+
+# For IQ-Tree trees (with branch lengths)
+p = ggplot(data = bs_df, aes(x = boxplot_fac, y = edge_length)) + geom_boxplot() +
+  facet_grid(dataset_fac ~ test_fac, scales = "free_y") + 
+  theme_bw() +
+  labs(title = "Concatenated trees: branch length of conflicted and congruent branches") +
+  scale_x_discrete(name = "Tree and edge type") +
+  scale_y_continuous(name = "Branch length (substitutions per site)") +
+  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 90, hjust = 1))
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_IQTree_branchLength_conflicting_branches.pdf"), plot = p, device = "pdf")
+ggsave(filename = paste0(node_output_dir, "plots/ShallowDatasets_IQTree_branchLength_conflicting_branches.png"), plot = p, device = "png")
+
+#### Plot the deep datasets ####
+# Separate out the two shallow datasets
+deep_df <- node_df[(node_df$dataset == "Strassert2021" | node_df$dataset == "1KP"),]
+
+# Separate into ASTRAL and IQ-Tree data frames
+pp_df <- deep_df[(deep_df$support_value_type == "PP"),]
+bs_df <- deep_df[(deep_df$support_value_type == "BS"),]
+
+# Make a nice plot of the deep datasets, faceted by test and dataset
+# For ASTRAL trees (with posterior probabilities)
+p = ggplot(data = pp_df, aes(x = boxplot_fac, y = support_value)) + geom_boxplot() +
+  facet_grid(dataset_fac ~ test_fac) + 
+  theme_bw() +
+  labs(title = "Coalescent trees: posterior probability of conflicted and congruent branches") +
+  scale_x_discrete(name = "Tree and edge type") +
+  scale_y_continuous(name = "Posterior Probability", breaks = seq(0,1,0.1),  labels = seq(0,1,0.1), minor_breaks = seq(0,1,0.05), limits = c(0,1)) +
+  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 90, hjust = 1))
+ggsave(filename = paste0(node_output_dir, "plots/DeepDatasets_ASTRAL_posteriorProbability_conflicting_branches.pdf"), plot = p, device = "pdf")
+ggsave(filename = paste0(node_output_dir, "plots/DeepDatasets_ASTRAL_posteriorProbability_conflicting_branches.png"), plot = p, device = "png")
+
+# For ASTRAL trees (with branch lengths)
+p = ggplot(data = pp_df, aes(x = boxplot_fac, y = edge_length)) + geom_boxplot() +
+  facet_grid(dataset_fac ~ test_fac) + 
+  theme_bw() +
+  labs(title = "Coalescent trees: branch length of conflicted and congruent branches") +
+  scale_x_discrete(name = "Tree and edge type") +
+  scale_y_continuous(name = "Branch length (coalescent units)", breaks = seq(0,6,2),  labels = seq(0,6,2), minor_breaks = seq(0,7,0.5), limits = c(0,7)) +
+  theme(plot.title = element_text(hjust = 0.5), axis.text.x = element_text(angle = 90, hjust = 1))
+ggsave(filename = paste0(node_output_dir, "plots/DeepDatasets_ASTRAL_branchLength_conflicting_branches.pdf"), plot = p, device = "pdf")
+ggsave(filename = paste0(node_output_dir, "plots/DeepDatasets_ASTRAL_branchLength_conflicting_branches.png"), plot = p, device = "png")
